src/main/java/com/Library/demo/controller/BorrowingRecordController.java
package com.Library.demo.controller;


import com.Library.demo.entity.BorrowingRecord;
import com.Library.demo.service.BorrowingRecordService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api")
public class BorrowingRecordController {

    @Autowired
    private BorrowingRecordService recordService;

    @PostMapping("/{bookId}/patron/{patronId}")
    public ResponseEntity<BorrowingRecord> borrowBook(@PathVariable Long bookId, @PathVariable Long patronId) {
        BorrowingRecord record = recordService.borrowBook(bookId, patronId);
        return ResponseEntity.status(HttpStatus.CREATED).body(record);
    }

    @PutMapping("/return/{bookId}/patron/{patronId}")
    public ResponseEntity<BorrowingRecord> returnBook(@PathVariable Long bookId, @PathVariable Long patronId) {
        BorrowingRecord record = recordService.returnBook(bookId, patronId);
        return ResponseEntity.ok(record);
    }
}
